src/main/java/io/saidmlx/actions/HelloAction.java
package io.saidmlx.actions;

import com.opensymphony.xwork2.ActionSupport;

/**
 * <code>Set welcome message.</code>
 */
public class HelloAction extends ActionSupport {

    public String execute() throws Exception {
    	System.out.println("HelloAction :: execute()");
    	
    	System.out.println("HelloAction :: execute()");
    	setMessage("Hola");
        //setMessage(getText(MESSAGE));
        System.out.print("+");
        return SUCCESS;
    }

    /**
     * Provide default valuie for Message property.
     */
    public static final String MESSAGE = "hello.message";

    /**
     * Field for Message property.
     */
    private String message;

    /**
     * Return Message property.
     *
     * @return Message property
     */
    public String getMessage() {
        return message;
    }

    /**
     * Set Message property.
     *
     * @param message Text to display on HelloWorld page.
     */
    public void setMessage(String message) {
        this.message = message;
    }
}
